pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-search-api"
version = "2.0.0"
description = "AI-Powered Document Search System with Flexible Architecture"
authors = [
    {name = "Bui-Tung-Hung", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "qdrant-client>=1.7.0",
    "aiohttp>=3.9.1",
    "numpy>=1.24.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/Bui-Tung-Hung/The-Documents-Manager"
Repository = "https://github.com/Bui-Tung-Hung/The-Documents-Manager"
Documentation = "https://github.com/Bui-Tung-Hung/The-Documents-Manager#readme"
"Bug Tracker" = "https://github.com/Bui-Tung-Hung/The-Documents-Manager/issues"

[project.scripts]
doc-search-api = "app.main:main"
check-config = "tools.check_config:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "tools*"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "qdrant_client.*"
ignore_missing_imports = true
